+++ CHANGELOG.md
@@ -4,6 +4,7 @@ Changelog for Flamenco Manager
 ## Version 2.0.9 (released 2017-05-11)
 
 - Workers: only store host part of worker's address, and not the port number.
+- Dashboard: Shorten task and worker IDs and nicer timestamp formatting.
 
 
 ## Version 2.0.8 (released 2017-05-09)